feat(area-theme): MigrationFlow を area から theme へ責務移管
人口移動フローは「area = 県固有ストーリー」ではなく「theme = 指標軸×横断
可視化」に属するため (docs/01_技術設計/11_area-vs-theme責務分離.md)、
/areas/[areaCode] から削除し /themes/population-dynamics 配下に移管。

- AreaMigrationFlowSection を削除 (areas/[areaCode]/page.tsx, index.ts)
- ThemeMigrationFlowSection を新設 (showSelector=true で県切替可、IO 遅延)
- ThemePageLayout で theme.themeKey === "population-dynamics" 時に埋め込み

棚卸し (`page-components-audit.cjs`) 結果: pageType=area の 92 件中 DB
レベルの違反 0、本件は page_components ではなくコード直書きの責務違反だった。

検証: next build で /areas/[areaCode]=● SSG, /themes/population-dynamics
=○ Static を維持 (SSG-preservation OK)。

棚卸し結果: docs/04_レビュー/area-theme-audit/2026-05-27.md
